LPS-154326 Move Util class to api
